Move RemoteSessionRepo into SessionRepository
