+++ configserver/src/main/java/com/yahoo/vespa/config/server/tenant/Tenant.java
@@ -7,7 +7,6 @@ import com.yahoo.vespa.config.server.ReloadHandler;
 import com.yahoo.vespa.config.server.RequestHandler;
 import com.yahoo.vespa.config.server.application.TenantApplications;
 import com.yahoo.vespa.config.server.session.SessionRepository;
-import com.yahoo.vespa.config.server.session.RemoteSessionRepo;
 import com.yahoo.vespa.config.server.session.SessionFactory;
 import com.yahoo.vespa.curator.Curator;
 import org.apache.zookeeper.data.Stat;
@@ -28,7 +27,6 @@ public class Tenant implements TenantHandlerProvider {
     static final String APPLICATIONS = "applications";
 
     private final TenantName name;
-    private final RemoteSessionRepo remoteSessionRepo;
     private final Path path;
     private final SessionFactory sessionFactory;
     private final SessionRepository sessionRepository;
@@ -40,7 +38,6 @@ public class Tenant implements TenantHandlerProvider {
     Tenant(TenantName name,
            SessionFactory sessionFactory,
            SessionRepository sessionRepository,
-           RemoteSessionRepo remoteSessionRepo,
            RequestHandler requestHandler,
            ReloadHandler reloadHandler,
            TenantApplications applicationRepo,
@@ -49,7 +46,6 @@ public class Tenant implements TenantHandlerProvider {
         this.path = TenantRepository.getTenantPath(name);
         this.requestHandler = requestHandler;
         this.reloadHandler = reloadHandler;
-        this.remoteSessionRepo = remoteSessionRepo;
         this.sessionFactory = sessionFactory;
         this.sessionRepository = sessionRepository;
         this.applicationRepo = applicationRepo;
@@ -74,13 +70,8 @@ public class Tenant implements TenantHandlerProvider {
         return requestHandler;
     }
 
-    /**
-     * The RemoteSessionRepo for this
-     *
-     * @return repo
-     */
-    public RemoteSessionRepo getRemoteSessionRepo() {
-        return remoteSessionRepo;
+    public SessionRepository getSessionRepo() {
+        return sessionRepository;
     }
 
     public TenantName getName() {
@@ -140,7 +131,7 @@ public class Tenant implements TenantHandlerProvider {
      * Called by watchers as a reaction to {@link #delete()}.
      */
     void close() {
-        remoteSessionRepo.close();              // Closes watchers and clears memory.
+        sessionRepository.close();              // Closes watchers and clears memory.
         applicationRepo.close();                // Closes watchers.
         sessionRepository.close();               // Closes watchers, clears memory, and deletes local files and ZK session state.
     }
